util: concurrency: Move concurrency primitives to `util` crate
