+++ common/src/lib.rs
@@ -9,26 +9,7 @@
 #![deny(clippy::missing_docs_in_private_items)]
 #![feature(generic_const_exprs)]
 
-use std::sync::{Arc, RwLock};
-use tokio::sync::RwLock as TokioRwLock;
-
 pub mod default_wrapper;
 pub mod keyed_list;
 pub mod types;
 pub mod worker;
-
-/// A type alias for a shared, concurrency safe, mutable pointer
-pub type Shared<T> = Arc<RwLock<T>>;
-/// A type alias for a shared, concurrency safe, mutable pointer in an
-/// async context
-pub type AsyncShared<T> = Arc<TokioRwLock<T>>;
-
-/// Wrap an abstract value in a shared lock
-pub fn new_shared<T>(wrapped: T) -> Shared<T> {
-    Arc::new(RwLock::new(wrapped))
-}
-
-/// Wrap an abstract value in an async shared lock
-pub fn new_async_shared<T>(wrapped: T) -> AsyncShared<T> {
-    Arc::new(TokioRwLock::new(wrapped))
-}
