+++ util/src/lib.rs
@@ -9,12 +9,12 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
 pub mod arbitrum;
+pub mod concurrency;
 pub mod errors;
 pub mod hex;
 pub mod matching_engine;
 pub mod metered_channels;
 pub mod networking;
-pub mod runtime;
 pub mod telemetry;
 
 /// Returns the current unix timestamp in seconds, represented as u64
